feat: implement upgrade, installation, versions and schematics
Make Talos not installed initially.
Then when the config is applied read the machine config and extract
version, schematic and install disk from the install section.
Validate install disk and mark picked disk as the system disk.
Update Talos version and Schematic resources in the state.

Implement `MachineStatus` controller that reacts on Version and
Schematic presence. Set `MachineStatus` as running and ready when both
resources exist.

Implement `Upgrade` API which extracts schematic and Talos version from
the image and immediately updates it in the state.

Signed-off-by: Artem Chernyshev <artem.chernyshev@talos-systems.com>