feat: support partial machine configs
Do not panic when the applied config is partial - handle it correctly.

Add a unit test for it.

Split some files using Linux-specific APIs and constants to be able to compile and run unit tests on non-Linux systems (e.g., MacOS).

Add a panic handler to the GRPC server to not crash the emulator on panics.

Bump Go & rekres.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>