feat: add schematic id to schematic conversion and cmdline support
Add a new service to convert schematic IDs to raw schematics. It uses a cache to not download initramfs for the schematics which were already seen.

Use this service to compute the `ExtensionStatus` resources in a more realistic manner.
Also use this service to produce a `KernelCmdline` resource.

Remove the "known schematics" concept, since we do not need that anymore.

Add more replace directives to `go.mod` to make Goland happy.

Signed-off-by: Utku Ozdemir <utku.ozdemir@siderolabs.com>